python-app/requirements.txt
fastapi==0.104.0
inngest==0.3.0a1
python-dotenv==1.0.0 
uvicorn==0.23.2
